move customized layers, call backs to new .py file 'clayers.py'. remove activation from equal conv/dense fn.
